requirements.txt
apache-airflow==2.5.0
apache-airflow-providers-postgres==5.4.0
pandas==1.5.2
elasticsearch==8.6.0